# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hospital operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  name: string;
  doctorCount: number;
}

export interface DepartmentInput {
  /** @minLength 1 */
  name: string;
}

export interface Doctor {
  id: number;
  name: string;
  departmentId: number;
  departmentName: string;
}

export interface DoctorInput {
  /** @minLength 1 */
  name: string;
  departmentId: number;
}

export interface Patient {
  id: number;
  name: string;
  registrationDate: string;
  appointmentCount: number;
}

export interface PatientInput {
  /** @minLength 1 */
  name: string;
  registrationDate: string;
}

export interface Appointment {
  id: number;
  appointmentDate: string;
  description: string;
  patientId: number;
  patientName: string;
  doctorId: number;
  doctorName: string;
  departmentName: string;
}

export interface AppointmentInput {
  appointmentDate: string;
  /** @minLength 1 */
  description: string;
  patientId: number;
  doctorId: number;
}

export interface Dashboard {
  departmentCount: number;
  doctorCount: number;
  patientCount: number;
  appointmentCount: number;
  todayAppointments: Appointment[];
  upcomingAppointments: Appointment[];
}

export type NotFoundResponse = {
  error: string;
};