# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospital operations API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Hospital overview
  - name: departments
  - name: doctors
  - name: patients
  - name: appointments
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get hospital dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Department" }
    post:
      operationId: createDepartment
      tags: [departments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DepartmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateDepartment
      tags: [departments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DepartmentInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
    delete:
      operationId: deleteDepartment
      tags: [departments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "204": { description: Deleted }
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      responses:
        "200":
          description: Doctors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Doctor" }
    post:
      operationId: createDoctor
      tags: [doctors]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DoctorInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
    patch:
      operationId: updateDoctor
      tags: [doctors]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DoctorInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
    delete:
      operationId: deleteDoctor
      tags: [doctors]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "204": { description: Deleted }
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      responses:
        "200":
          description: Patients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Patient" }
    post:
      operationId: createPatient
      tags: [patients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatientInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
    patch:
      operationId: updatePatient
      tags: [patients]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatientInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
    delete:
      operationId: deletePatient
      tags: [patients]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "204": { description: Deleted }
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
    post:
      operationId: createAppointment
      tags: [appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
    patch:
      operationId: updateAppointment
      tags: [appointments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "204": { description: Deleted }
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: number }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
            required: [error]
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Department:
      type: object
      required: [id, name, doctorCount]
      properties:
        id: { type: number }
        name: { type: string }
        doctorCount: { type: number }
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    Doctor:
      type: object
      required: [id, name, departmentId, departmentName]
      properties:
        id: { type: number }
        name: { type: string }
        departmentId: { type: number }
        departmentName: { type: string }
    DoctorInput:
      type: object
      required: [name, departmentId]
      properties:
        name: { type: string, minLength: 1 }
        departmentId: { type: number }
    Patient:
      type: object
      required: [id, name, registrationDate, appointmentCount]
      properties:
        id: { type: number }
        name: { type: string }
        registrationDate: { type: string, format: date }
        appointmentCount: { type: number }
    PatientInput:
      type: object
      required: [name, registrationDate]
      properties:
        name: { type: string, minLength: 1 }
        registrationDate: { type: string, format: date }
    Appointment:
      type: object
      required: [id, appointmentDate, description, patientId, patientName, doctorId, doctorName, departmentName]
      properties:
        id: { type: number }
        appointmentDate: { type: string, format: date-time }
        description: { type: string }
        patientId: { type: number }
        patientName: { type: string }
        doctorId: { type: number }
        doctorName: { type: string }
        departmentName: { type: string }
    AppointmentInput:
      type: object
      required: [appointmentDate, description, patientId, doctorId]
      properties:
        appointmentDate: { type: string, format: date-time }
        description: { type: string, minLength: 1 }
        patientId: { type: number }
        doctorId: { type: number }
    Dashboard:
      type: object
      required: [departmentCount, doctorCount, patientCount, appointmentCount, todayAppointments, upcomingAppointments]
      properties:
        departmentCount: { type: number }
        doctorCount: { type: number }
        patientCount: { type: number }
        appointmentCount: { type: number }
        todayAppointments:
          type: array
          items: { $ref: "#/components/schemas/Appointment" }
        upcomingAppointments:
          type: array
          items: { $ref: "#/components/schemas/Appointment" }