# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hospital operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Appointment {
  id: number;
  appointmentDate: Date;
  description: string;
  patientId: number;
  patientName: string;
  doctorId: number;
  doctorName: string;
  departmentName: string;
}