# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hospital operations API
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  appointmentDate: Date;
  /** @minLength 1 */
  description: string;
  patientId: number;
  doctorId: number;
}