# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hospital operations API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';

export interface Dashboard {
  departmentCount: number;
  doctorCount: number;
  patientCount: number;
  appointmentCount: number;
  todayAppointments: Appointment[];
  upcomingAppointments: Appointment[];
}