# lib/api-zod/src/generated/types/department.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hospital operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Department {
  id: number;
  name: string;
  doctorCount: number;
}