# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hospital operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  departmentId: number;
  departmentName: string;
}